+++ braintree/transaction_line_item_gateway.py
@@ -4,7 +4,7 @@ from braintree.resource import Resource
 from braintree.resource_collection import ResourceCollection
 from braintree.transaction_line_item import TransactionLineItem
 from braintree.exceptions.not_found_error import NotFoundError
-from braintree.exceptions.down_for_maintenance_error import DownForMaintenanceError
+from braintree.exceptions.timeout_error import TimeoutError
 
 class TransactionLineItemGateway(object):
     def __init__(self, gateway):
@@ -18,8 +18,7 @@ class TransactionLineItemGateway(object):
             response = self.config.http().get(self.config.base_merchant_path() + "/transactions/" + transaction_id + "/line_items")
             if "line_items" in response:
                 return [TransactionLineItem(item) for item in ResourceCollection._extract_as_array(response, "line_items")]
-            # TODO ServiceUnavailable or UnknownException?
             else:
-                raise DownForMaintenanceError()
+                raise TimeoutError()
         except NotFoundError:
             raise NotFoundError("transaction line items with id " + repr(transaction_id) + " not found")